добавил конвертеры (RoomConverters (для ROOM) и ExtendedIngredientConverter).
подключил библиотеку GSON.

Сконвертировал список ингридиентов и передал их в ROOM.
изменил тип Int на Long у переменной Id.

+++ app/src/main/java/com/agobikk/cookeatenjoy/model/Models.kt
@@ -37,7 +37,7 @@ data class ResultMainCourse(
 
 data class FoodInformation(
     @SerializedName("id")
-    val id: Int,
+    val id: Long,
     @SerializedName("image")
     val image: String,
     @SerializedName("summary")
@@ -53,7 +53,7 @@ data class FoodInformation(
 
 data class ExtendedIngredient(
     @SerializedName("id")
-    val idExtendedIngredient: Int,
+    val idExtendedIngredient: Long,
     @SerializedName("amount")
     val amount: Double?,
     @SerializedName("consistency")
